UWP: Move some functions to a more sensible place
